move back the class with `@ConfigMapping`
