+++ src/main/java/io/stargate/sgv2/jsonapi/api/v1/metrics/JsonApiMetricsConfig.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.metrics;
+package io.stargate.sgv2.jsonapi.api.v1.metrics;
 
 import io.smallrye.config.ConfigMapping;
 import io.smallrye.config.WithDefault;
